+++ train_node.py
@@ -84,10 +84,10 @@ if not os.path.isfile('embs/' + emb_file_name):
     optimizer = torch.optim.Adam(model.parameters(), lr=train_param['lr'])
     sampler = None
     if not ('no_sample' in sample_param and sample_param['no_sample']):
-        sampler = ParallelSampler(g['indptr'], g['indices'], g['eid'], g['ts'].astype(np.float32),
-                                sample_param['num_thread'], 1, sample_param['layer'], sample_param['neighbor'],
-                                sample_param['strategy']=='recent', sample_param['prop_time'],
-                                sample_param['history'], float(sample_param['duration']))
+        sampler = ParallelSampler(g['indptr'], g['indices'], g['eid'], g['ts'].astype(np.int64),
+                                  sample_param['num_thread'], 1, sample_param['layer'], sample_param['neighbor'],
+                                  sample_param['strategy']=='recent', sample_param['prop_time'],
+                                  sample_param['history'], float(sample_param['duration']))
     neg_link_sampler = NegLinkSampler(g['indptr'].shape[0] - 1)
 
     model.load_state_dict(torch.load(args.model))
@@ -105,7 +105,7 @@ if not os.path.isfile('embs/' + emb_file_name):
             else:
                 model.eval()
             root_nodes = np.concatenate([rows.src.values, rows.dst.values, neg_link_sampler.sample(len(rows))]).astype(np.int32)
-            ts = np.concatenate([rows.time.values, rows.time.values, rows.time.values]).astype(np.float32)
+            ts = np.concatenate([rows.time.values, rows.time.values, rows.time.values]).astype(np.int64)
             if sampler is not None:
                 if 'no_neg' in sample_param and sample_param['no_neg']:
                     pos_root_end = root_nodes.shape[0] * 2 // 3
@@ -156,7 +156,7 @@ if not os.path.isfile('embs/' + emb_file_name):
         total=(len(ldf) + args.batch_size - 1) // args.batch_size,
     ) if args.tqdm else ldf.groupby(ldf.index // args.batch_size)
     for _, rows in emb_iter:
-        emb.append(get_node_emb(rows.node.values.astype(np.int32), rows.time.values.astype(np.float32)))
+        emb.append(get_node_emb(rows.node.values.astype(np.int32), rows.time.values.astype(np.int64)))
     emb = torch.cat(emb, dim=0)
     torch.save(emb, 'embs/' + emb_file_name)
     print('Saved to embs/' + emb_file_name)
